Change prefix meaning of evil-org-insert-subheading/subtodo
The old meaning is to insert a heading and to copy the heading text n - 1 times
on the lines below that heading.
Instead it now follows the same prefix argument logic as org-insert-heading.
If one prefix argument is given, the insertion will be done at the end of the
current subtree.