Bind calendar keys directly to the Emacs state, not to Evil insert state
The minibuffer is not running Evil mode by default.  The extra configuration it
requires might be a hassle we don't want to tackle here.

Setting the bindings directly on the map will probably not interfere anyways as
the default values of those keys are not very useful to select a date.